requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
et_xmlfile==2.0.0
h11==0.16.0
idna==3.10
openpyxl==3.1.5
outcome==1.3.0.post0
pillow==11.2.1
pycparser==2.22
PySocks==1.7.1
selenium==4.33.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
